Move AttributeDictionary to global util file, because now it is also imported in context.py
